+++ lib/src/utils/test_ad_id_manager.dart
@@ -29,13 +29,16 @@ class TestAdIdManager extends IAdIdManager {
 
   @override
   AppAdIds? get appLovinAdIds => AppAdIds(
-        appId:
-            'OeKTS4Zl758OIlAs3KQ6-3WE1IkdOo3nQNJtRubTzlyFU76TRWeQZAeaSMCr9GcZdxR4p2cnoZ1Gg7p7eSXCdA',
-        bannerId: Platform.isAndroid ? 'b2c4f43d3986bcfb' : '80c269494c0e45c2',
-        interstitialId:
-            Platform.isAndroid ? 'c48f54c6ce5ff297' : 'e33147110a6d12d2',
-        rewardedId:
-            Platform.isAndroid ? 'ffbed216d19efb09' : 'f4af3e10dd48ee4f',
+        appId: 'YOUR_SDK_KEY',
+        bannerId: Platform.isAndroid
+            ? 'ANDROID_BANNER_AD_UNIT_ID'
+            : 'IOS_BANNER_AD_UNIT_ID',
+        interstitialId: Platform.isAndroid
+            ? 'ANDROID_INTER_AD_UNIT_ID'
+            : 'IOS_INTER_AD_UNIT_ID',
+        rewardedId: Platform.isAndroid
+            ? 'ANDROID_REWARDED_AD_UNIT_ID'
+            : 'IOS_REWARDED_AD_UNIT_ID',
       );
 
   @override